Refactor doubletake to have basic structure
