+++ frontend/public/css/icons.css
@@ -199,3 +199,6 @@
     margin-bottom: 0px;
   }
 };
+
+/* doubletake */
+/* undo */
