+++ frontend/public/index.html
@@ -21,6 +21,7 @@
     <link rel="stylesheet" href="./css/modal_form.css" />
     <link rel="stylesheet" href="./css/attributes_form.css" />
     <link rel="stylesheet" href="./css/essay_form.css" />
+    <link rel="stylesheet" href="./css/doubletake.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
